Refactor GameView layout and observers into dedicated extensions
