+++ UI/GameView.swift
@@ -138,7 +138,7 @@ struct GameView: View {
             GameMenuActionConfirmationSheet(
                 action: action,
                 onConfirm: { confirmedAction in
-                    // performMenuAction 内でも viewModel.pendingMenuAction を破棄しているが、
+                    // performMenuAction 内で viewModel.pendingMenuAction を破棄しているが、
                     // 明示的に nil を代入しておくことでバインディング由来のシート閉鎖と状態初期化を二重に保証する
                     performMenuAction(confirmedAction)
                     viewModel.pendingMenuAction = nil
@@ -178,333 +178,6 @@ struct GameView: View {
         }
     }
 
-    /// GeometryReader 内部のレイアウト調整と描画処理をまとめたメインコンテンツ
-    /// - Parameter geometry: 外側から渡される GeometryProxy（親ビューのサイズや安全領域を把握するために利用）
-    /// - Returns: レイアウト計算結果を反映したゲームプレイ画面全体のビュー階層
-    @ViewBuilder
-    private func mainContent(for geometry: GeometryProxy) -> some View {
-        // MARK: - レイアウト関連の計算結果を専用コンテキストへ集約
-        // 単一メソッドで値を求めておくことで ViewBuilder の複雑さを抑え、コンパイラの型推論負荷を軽減する。
-        let layoutCalculator = GameViewLayoutCalculator(
-            geometry: geometry,
-            horizontalSizeClass: horizontalSizeClass,
-            topOverlayHeight: topOverlayHeight,
-            baseTopSafeAreaInset: baseTopSafeAreaInset,
-            statisticsHeight: viewModel.statisticsHeight,
-            handSectionHeight: viewModel.handSectionHeight
-        )
-        let layoutContext = layoutCalculator.makeContext()
-        // 監視用の不可視オーバーレイも先に生成し、View ビルダー内でのネストを浅く保つ
-        let diagnosticsOverlay = layoutDiagnosticOverlay(using: layoutContext)
-
-        ZStack(alignment: .top) {
-            VStack(spacing: GameViewLayoutMetrics.spacingBetweenBoardAndHand) {
-                boardSection(width: layoutContext.boardWidth)
-                GameHandSectionView(
-                    theme: theme,
-                    viewModel: viewModel,
-                    boardBridge: boardBridge,
-                    cardAnimationNamespace: cardAnimationNamespace,
-                    handSlotCount: handSlotCount,
-                    bottomInset: layoutContext.bottomInset,
-                    bottomPadding: layoutContext.handSectionBottomPadding
-                )
-            }
-            // 統計バッジ＋操作ボタンを上部へ寄せ、ノッチやステータスバーと干渉しないように余白を加算
-            .padding(.top, layoutContext.controlRowTopPadding)
-            // MARK: - 手詰まりペナルティ通知バナー
-            penaltyBannerOverlay(contentTopInset: layoutContext.overlayAdjustedTopInset)
-        }
-        // 画面全体の背景もテーマで制御し、システム設定と調和させる
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .background(theme.backgroundPrimary)
-        // 盤面が表示されない不具合を切り分けるため、レイアウト関連の値をウォッチする不可視ビューを重ねる
-        .background(diagnosticsOverlay)
-    }
-
-    /// PreferenceKey や Combine パブリッシャによる状態監視モディファイアをひとまとめに適用する
-    /// - Parameter content: 監視対象となるベースビュー
-    /// - Returns: 各種監視モディファイアを適用済みのビュー
-    private func applyGameViewObservers<Content: View>(to content: Content) -> some View {
-        content
-        // PreferenceKey で伝搬した各セクションの高さを受け取り、レイアウト計算に利用する
-            .onPreferenceChange(StatisticsHeightPreferenceKey.self) { newHeight in
-                // 統計バッジ領域の計測値が変わったタイミングで旧値と比較し、変化量をログへ残して原因調査を容易にする
-                let previousHeight = viewModel.statisticsHeight
-                guard previousHeight != newHeight else { return }
-                debugLog("GameView.viewModel.statisticsHeight 更新: 旧値=\(previousHeight), 新値=\(newHeight)")
-                viewModel.statisticsHeight = newHeight
-            }
-            .onPreferenceChange(HandSectionHeightPreferenceKey.self) { newHeight in
-                // 手札セクションの高さ変化も逐次観測し、ホームインジケータ付近での余白不足を切り分けられるようにする
-                let previousHeight = viewModel.handSectionHeight
-                guard previousHeight != newHeight else { return }
-                debugLog("GameView.viewModel.handSectionHeight 更新: 旧値=\(previousHeight), 新値=\(newHeight)")
-                viewModel.handSectionHeight = newHeight
-            }
-            // 盤面 SpriteView のアンカー更新を監視し、アニメーションの移動先として保持
-            .onPreferenceChange(BoardAnchorPreferenceKey.self) { anchor in
-                viewModel.updateBoardAnchor(anchor)
-            }
-            // 初回表示時に SpriteKit の背景色もテーマに合わせて更新
-            .onAppear {
-                viewModel.prepareForAppear(
-                    colorScheme: colorScheme,
-                    guideModeEnabled: guideModeEnabled,
-                    hapticsEnabled: hapticsEnabled,
-                    handOrderingStrategy: resolveHandOrderingStrategy()
-                )
-            }
-            // ライト/ダーク切り替えが発生した場合も SpriteKit 側へ反映
-            .onChange(of: colorScheme) { _, newScheme in
-                viewModel.applyScenePalette(for: newScheme)
-                // カラースキーム変更時はガイドの色味も再描画して視認性を確保
-                viewModel.refreshGuideHighlights()
-            }
-            // 手札の並び設定が変わったら即座にゲームロジックへ伝え、UI の並びも更新する
-            .onChange(of: handOrderingRawValue) { _, _ in
-                viewModel.applyHandOrderingStrategy(rawValue: handOrderingRawValue)
-            }
-            // ガイドモードのオン/オフを切り替えたら即座に ViewModel 側へ委譲する
-            .onChange(of: guideModeEnabled) { _, isEnabled in
-                viewModel.updateGuideMode(enabled: isEnabled)
-            }
-            // ハプティクス設定が切り替わった際も ViewModel へ伝え、サービス呼び出しを統一する
-            .onChange(of: hapticsEnabled) { _, isEnabled in
-                viewModel.updateHapticsSetting(isEnabled: isEnabled)
-            }
-            // 経過時間を 1 秒ごとに再計算し、リアルタイム表示へ反映
-            .onReceive(viewModel.elapsedTimer) { _ in
-                viewModel.updateDisplayedElapsedTime()
-            }
-            // カードが盤面へ移動中は UI 全体を操作不可とし、状態の齟齬を防ぐ
-            .disabled(boardBridge.animatingCard != nil)
-            // Preference から取得したアンカー情報を用いて、カードが盤面中央へ吸い込まれる演出を重ねる
-            .overlayPreferenceValue(CardPositionPreferenceKey.self) { anchors in
-                GameCardAnimationOverlay(
-                    anchors: anchors,
-                    boardBridge: boardBridge,
-                    fallbackCurrentPosition: viewModel.currentPosition,
-                    cardAnimationNamespace: cardAnimationNamespace
-                )
-            }
-    }
-
-    /// レギュラー幅（iPad）向けにシート表示へ切り替えるためのバインディング
-    /// - Returns: iPad では viewModel.pendingMenuAction を返し、それ以外では常に nil を返すバインディング
-    private var regularWidthPendingActionBinding: Binding<GameMenuAction?> {
-        Binding(
-            get: {
-                // 横幅が十分でない場合はシート表示を抑制し、確認ダイアログ側に処理を委ねる
-                guard horizontalSizeClass == .regular else {
-                    return nil
-                }
-                return viewModel.pendingMenuAction
-            },
-            set: { newValue in
-                // シートを閉じたときに SwiftUI から nil が渡されるため、そのまま状態へ反映しておく
-                viewModel.pendingMenuAction = newValue
-            }
-        )
-    }
-
-    /// 盤面の統計と SpriteKit ボードをまとめて描画する
-    /// - Parameter width: GeometryReader で算出した盤面の幅（正方形表示の基準）
-    /// - Returns: 統計バッジと SpriteView を縦に並べた領域
-    private func boardSection(width: CGFloat) -> some View {
-        VStack(alignment: .leading, spacing: GameViewLayoutMetrics.spacingBetweenStatisticsAndBoard) {
-            GameBoardControlRowView(
-                theme: theme,
-                viewModel: viewModel
-            )
-            ZStack {
-                spriteBoard(width: width)
-                if viewModel.progress == .awaitingSpawn {
-                    spawnSelectionOverlay
-                        // 盤面いっぱいに広がりすぎないよう最大幅を制限する
-                        .frame(maxWidth: width * 0.9)
-                        .padding(.horizontal, 12)
-                }
-            }
-            // 盤面縮小で生まれた余白を均等にするため、中央寄せで描画する
-            .frame(maxWidth: .infinity, alignment: .center)
-        }
-    }
-
-    /// SpriteKit の盤面を描画し、ライフサイクルに応じた更新処理をまとめる
-    /// - Parameter width: 正方形に保つための辺長
-    /// - Returns: onAppear / onReceive を含んだ SpriteView
-    private func spriteBoard(width: CGFloat) -> some View {
-        SpriteView(scene: scene)
-            // 正方形で表示したいため幅に合わせる
-            .frame(width: width, height: width)
-            // 盤面のアンカーを収集し、カード移動アニメーションの着地点に利用
-            .anchorPreference(key: BoardAnchorPreferenceKey.self, value: .bounds) { $0 }
-            .onAppear {
-                // BoardBridge 側で SpriteKit シーンと GameCore の同期をまとめて実施
-                boardBridge.configureSceneOnAppear(width: width)
-            }
-            // ジオメトリの変化に追従できるよう、SpriteKit シーンのサイズも都度更新する
-            .onChange(of: width) { _, newWidth in
-                boardBridge.updateSceneSize(to: newWidth)
-            }
-    }
-
-    /// スポーン位置選択中に盤面へ重ねる案内オーバーレイ
-    private var spawnSelectionOverlay: some View {
-        VStack(spacing: 8) {
-            Text("開始マスを選択")
-                .font(.system(size: 18, weight: .semibold, design: .rounded))
-            Text("手札スロットと先読みを確認してから、好きなマスをタップしてください。")
-                .font(.system(size: 14, weight: .medium, design: .rounded))
-                .multilineTextAlignment(.center)
-        }
-        .padding(.vertical, 18)
-        .padding(.horizontal, 20)
-        .background(
-            RoundedRectangle(cornerRadius: 16, style: .continuous)
-                .fill(theme.spawnOverlayBackground)
-                .overlay(
-                    RoundedRectangle(cornerRadius: 16, style: .continuous)
-                        .stroke(theme.spawnOverlayBorder, lineWidth: 1)
-                )
-        )
-        .shadow(color: theme.spawnOverlayShadow, radius: 18, x: 0, y: 8)
-        .foregroundColor(theme.textPrimary)
-        .allowsHitTesting(false)
-        .accessibilityElement(children: .combine)
-        .accessibilityLabel(Text("開始位置を選択してください。手札スロットと次のカードを見てから任意のマスをタップできます。"))
-    }
-
-    /// 手詰まりペナルティを知らせるバナーのレイヤーを構成
-    private func penaltyBannerOverlay(contentTopInset: CGFloat) -> some View {
-        // MARK: - ステータスバーとの距離を安全に確保する
-        // iPad のフォームシートなどで safeAreaInsets.top が 0 になるケースでは、
-        // バナーが画面最上部へ貼り付いてしまうため、フォールバックを交えつつ余白を広げる。
-        // contentTopInset にはステータスバー由来の安全領域のみを渡し、RootView のトップバーぶんはすでに差し引いた状態にする。
-        // これによりトップバー表示時でもバナーが極端に下へずり落ちることを防ぐ。
-        // contentTopInset が 0 でも GameViewLayoutMetrics.penaltyBannerBaseTopPadding だけは必ず確保し、
-        // 非ゼロのインセットが得られた場合は追加マージンを加えて矢印付きダイアログとの干渉を避ける。
-        let resolvedTopPadding = max(
-            GameViewLayoutMetrics.penaltyBannerBaseTopPadding,
-            contentTopInset + GameViewLayoutMetrics.penaltyBannerSafeAreaAdditionalPadding
-        )
-
-        return VStack {
-            if viewModel.isShowingPenaltyBanner {
-                HStack {
-                    Spacer(minLength: 0)
-                    PenaltyBannerView(penaltyAmount: viewModel.lastPenaltyAmount)
-                        .padding(.horizontal, 20)
-                        .transition(.move(edge: .top).combined(with: .opacity))
-                        .accessibilityIdentifier("penalty_banner")
-                    Spacer(minLength: 0)
-                }
-            }
-            Spacer()
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .padding(.top, resolvedTopPadding)
-        .allowsHitTesting(false)  // バナーが表示されていても下の UI を操作可能にする
-        .zIndex(2)
-    }
-
-    /// レイアウトに関する最新の実測値をログに残すための不可視ビューを生成
-    /// - Parameter context: GeometryReader から抽出したレイアウト情報コンテキスト
-    /// - Returns: 画面上には表示されない監視用ビュー
-    private func layoutDiagnosticOverlay(using context: GameViewLayoutContext) -> some View {
-        // 現在のレイアウト関連値をひとまとめにして Equatable なスナップショットとして扱い、差分が生じたときだけログを出力する
-        let snapshot = BoardLayoutSnapshot(context: context)
-
-        return Color.clear
-            .frame(width: 0, height: 0)
-            .allowsHitTesting(false)
-            .onAppear {
-                // 表示直後にレイアウト情報を記録して初期状態を把握
-                logLayoutSnapshot(snapshot, reason: "初期観測")
-            }
-            .onChange(of: snapshot) { _, newValue in
-                // レイアウト値が変動するたびにスナップショットを残し、問題の再現条件を追跡する
-                logLayoutSnapshot(newValue, reason: "値更新")
-            }
-    }
-
-    /// 実測値のスナップショットを比較しつつログへ出力する共通処理
-    /// - Parameters:
-    ///   - snapshot: 記録したいレイアウト値一式
-    ///   - reason: ログ出力の契機（onAppear / 値更新など）
-    private func logLayoutSnapshot(_ snapshot: BoardLayoutSnapshot, reason: String) {
-        // 同じ値での重複出力を避け、必要なタイミングのみに絞ってログを残す
-        if viewModel.lastLoggedLayoutSnapshot == snapshot { return }
-        viewModel.lastLoggedLayoutSnapshot = snapshot
-
-        // 盤面縮小ロジックのどこで値が想定外になっているか突き止められるよう、多段の詳細ログを整形して出力する
-        let message = """
-        GameView.layout 観測: 理由=\(reason)
-          geometry=\(snapshot.geometrySize)
-          safeArea(rawTop=\(snapshot.rawTopInset), baseTop=\(snapshot.baseTopSafeAreaInset), rawBottom=\(snapshot.rawBottomInset), resolvedTop=\(snapshot.resolvedTopInset), overlayAdjustedTop=\(snapshot.overlayAdjustedTopInset), resolvedBottom=\(snapshot.resolvedBottomInset), fallbackTop=\(snapshot.usedTopSafeAreaFallback), fallbackBottom=\(snapshot.usedBottomSafeAreaFallback), overlayTop=\(snapshot.topOverlayHeight))
-          sections(statistics=\(snapshot.statisticsHeight), resolvedStatistics=\(snapshot.resolvedStatisticsHeight), hand=\(snapshot.handSectionHeight), resolvedHand=\(snapshot.resolvedHandSectionHeight))
-          paddings(controlTop=\(snapshot.controlRowTopPadding), handBottom=\(snapshot.handSectionBottomPadding), regularExtra=\(snapshot.regularAdditionalBottomPadding))
-          fallbacks(statistics=\(snapshot.usedStatisticsFallback), hand=\(snapshot.usedHandSectionFallback), topSafeArea=\(snapshot.usedTopSafeAreaFallback), bottomSafeArea=\(snapshot.usedBottomSafeAreaFallback))
-          boardBases(horizontal=\(snapshot.horizontalBoardBase), vertical=\(snapshot.verticalBoardBase), resolved=\(snapshot.boardBaseSize)) availableHeight=\(snapshot.availableHeight) boardScale=\(GameViewLayoutMetrics.boardScale) boardWidth=\(snapshot.boardWidth)
-        """
-
-        debugLog(message)
-
-        if snapshot.availableHeight <= 0 || snapshot.boardWidth <= 0 {
-            // 盤面がゼロサイズになる条件を明確化するため、異常時は追加で警告ログを残す
-            debugLog(
-                "GameView.layout 警告: availableHeight=\(snapshot.availableHeight), horizontalBase=\(snapshot.horizontalBoardBase), verticalBase=\(snapshot.verticalBoardBase), boardBase=\(snapshot.boardBaseSize), boardWidth=\(snapshot.boardWidth)"
-            )
-        }
-    }
-
-}
-
-/// GameViewLayoutContext から BoardLayoutSnapshot を組み立てるための補助イニシャライザ
-/// - Note: GameView 内部のみで利用するためアクセスレベルは private extension とする
-private extension BoardLayoutSnapshot {
-    init(context: GameViewLayoutContext) {
-        // GeometryReader から取得した実測値を丸ごとコピーし、ViewModel からも参照できる形へ変換
-        self.init(
-            geometrySize: context.geometrySize,
-            availableHeight: context.availableHeightForBoard,
-            horizontalBoardBase: context.horizontalBoardBase,
-            verticalBoardBase: context.verticalBoardBase,
-            boardBaseSize: context.boardBaseSize,
-            boardWidth: context.boardWidth,
-            rawTopInset: context.rawTopInset,
-            rawBottomInset: context.rawBottomInset,
-            baseTopSafeAreaInset: context.baseTopSafeAreaInset,
-            resolvedTopInset: context.topInset,
-            overlayAdjustedTopInset: context.overlayAdjustedTopInset,
-            resolvedBottomInset: context.bottomInset,
-            statisticsHeight: context.statisticsHeight,
-            resolvedStatisticsHeight: context.resolvedStatisticsHeight,
-            handSectionHeight: context.handSectionHeight,
-            resolvedHandSectionHeight: context.resolvedHandSectionHeight,
-            regularAdditionalBottomPadding: context.regularAdditionalBottomPadding,
-            handSectionBottomPadding: context.handSectionBottomPadding,
-            usedTopSafeAreaFallback: context.usedTopFallback,
-            usedBottomSafeAreaFallback: context.usedBottomFallback,
-            usedStatisticsFallback: context.usedStatisticsFallback,
-            usedHandSectionFallback: context.usedHandSectionFallback,
-            controlRowTopPadding: context.controlRowTopPadding,
-            topOverlayHeight: context.topOverlayHeight
-        )
-    }
-}
-
-private extension GameView {
-    /// AppStorage から読み出した文字列を安全に列挙体へ変換する
-    /// - Returns: 有効な設定値。未知の値は従来方式へフォールバックする
-    func resolveHandOrderingStrategy() -> HandOrderingStrategy {
-        HandOrderingStrategy(rawValue: handOrderingRawValue) ?? .insertionOrder
-    }
-}
-
-// MARK: - コントロールバーの操作要素
-private extension GameView {
     /// メニュー操作を実際に実行する共通処理
     /// - Parameter action: ユーザーが選択した操作種別
     private func performMenuAction(_ action: GameMenuAction) {
